Persist standalone consensus review verdicts
