+++ crates/cli-sub-agent/src/review_cmd_multi.rs
@@ -1,15 +1,253 @@
 use anyhow::{Context, Result};
+use csa_config::{GlobalConfig, ProjectConfig};
+use csa_core::consensus::AgentResponse;
 use csa_core::types::ToolName;
 use tokio::task::JoinSet;
-use tracing::error;
+use tracing::{error, warn};
 
+use crate::cli::ReviewArgs;
+use crate::pipeline::resolve_effective_initial_response_timeout_for_tool;
+use crate::review_consensus::{
+    CLEAN, agreement_level, build_multi_reviewer_instruction, consensus_strategy_label,
+    consensus_verdict, parse_consensus_strategy, resolve_consensus,
+};
+use crate::review_routing::ReviewRoutingMetadata;
+
+use super::bug_class_pipeline::{
+    maybe_extract_recurring_bug_class_skills, resolve_review_iterations,
+};
+use super::execute::{compute_diff_fingerprint, execute_review_with_tier_filter};
 use super::output::ReviewerOutcome;
 use super::output::{
     GEMINI_AUTH_PROMPT_STATUS_REASON, persist_review_meta, persist_review_verdict,
+    print_reviewer_outcomes,
 };
-use super::result_handling::build_unavailable_reviewer_outcome;
+use super::prior_rounds::explicit_review_tool;
+use super::result_handling::{build_reviewer_outcome, build_unavailable_reviewer_outcome};
+use super::reviewers::resolve_multi_reviewer_pool;
 use csa_session::state::ReviewSessionMeta;
 use std::path::Path;
+
+pub(super) struct MultiReviewerReviewContext<'a> {
+    pub args: &'a ReviewArgs,
+    pub reviewers: usize,
+    pub tool: ToolName,
+    pub prompt: &'a str,
+    pub scope: &'a str,
+    pub project_root: &'a Path,
+    pub config: &'a Option<ProjectConfig>,
+    pub global_config: &'a GlobalConfig,
+    pub pre_session_hook: Option<csa_hooks::PreSessionHookInvocation>,
+    pub review_routing: ReviewRoutingMetadata,
+    pub review_model: Option<String>,
+    pub resolved_model_spec: Option<String>,
+    pub resolved_tier_name: Option<String>,
+    pub review_thinking: Option<String>,
+    pub stream_mode: csa_process::StreamMode,
+    pub idle_timeout_seconds: u64,
+    pub readonly_project_root: bool,
+    pub prior_rounds_section: Option<&'a str>,
+}
+
+pub(super) async fn run_multi_reviewer_review(ctx: MultiReviewerReviewContext<'_>) -> Result<i32> {
+    if ctx.args.fix {
+        anyhow::bail!("--fix is not supported when --reviewers > 1");
+    }
+    if ctx.args.session.is_some() {
+        anyhow::bail!("--session is only supported when --reviewers=1");
+    }
+
+    let consensus_strategy = parse_consensus_strategy(&ctx.args.consensus)?;
+    let reviewer_pool = resolve_multi_reviewer_pool(
+        ctx.reviewers,
+        explicit_review_tool(ctx.args),
+        ctx.tool,
+        ctx.resolved_tier_name.as_deref(),
+        ctx.config.as_ref(),
+        ctx.global_config,
+    )?;
+    let reviewer_tools = reviewer_pool.reviewer_tools;
+    let reviewer_tool_plan = reviewer_tools.clone();
+    let tier_reviewer_specs = reviewer_pool.tier_reviewer_specs;
+    warn_if_fast_mode_has_no_codex_reviewer(
+        ctx.args.fast_but_more_cost,
+        &reviewer_tool_plan,
+        &tier_reviewer_specs,
+    );
+
+    let mut join_set = JoinSet::new();
+    for (reviewer_index, reviewer_tool) in reviewer_tools.into_iter().enumerate() {
+        let reviewer_prompt = build_multi_reviewer_instruction(
+            ctx.prompt,
+            reviewer_index + 1,
+            reviewer_tool,
+            ctx.project_root,
+            ctx.prior_rounds_section,
+        );
+        let reviewer_model = ctx.review_model.clone();
+        let reviewer_project_root = ctx.project_root.to_path_buf();
+        let reviewer_config = ctx.config.as_ref().cloned();
+        let reviewer_global = ctx.global_config.clone();
+        let reviewer_pre_session_hook = ctx.pre_session_hook.clone();
+        let reviewer_description = format!(
+            "review[{}]: {}",
+            reviewer_index + 1,
+            crate::run_helpers::truncate_prompt(ctx.scope, 80)
+        );
+        let reviewer_routing = ctx.review_routing.clone();
+
+        let reviewer_force_override = ctx.args.force_override_user_config;
+        let reviewer_force_ignore_tier = ctx.args.force_ignore_tier_setting;
+        let reviewer_no_failover = ctx.args.no_failover;
+        let reviewer_fast_but_more_cost = ctx.args.fast_but_more_cost;
+        let reviewer_no_fs_sandbox = ctx.args.no_fs_sandbox;
+        let reviewer_extra_writable = ctx.args.extra_writable.clone();
+        let reviewer_extra_readable = ctx.args.extra_readable.clone();
+        // Keep every reviewer on the resolved tier when possible by binding
+        // each tool to its tier model spec. Fall back to the primary spec only
+        // when we only have a single tier-resolved reviewer tool.
+        let reviewer_model_spec = tier_reviewer_specs
+            .iter()
+            .find(|resolution| resolution.tool == reviewer_tool)
+            .map(|resolution| resolution.model_spec.clone())
+            .or_else(|| {
+                if reviewer_tool == ctx.tool {
+                    ctx.resolved_model_spec.clone()
+                } else {
+                    None
+                }
+            });
+        let reviewer_tier_name = ctx.resolved_tier_name.clone();
+        let reviewer_thinking = ctx.review_thinking.clone();
+        let reviewer_initial_response_timeout_seconds =
+            resolve_effective_initial_response_timeout_for_tool(
+                reviewer_config.as_ref(),
+                ctx.args.initial_response_timeout,
+                ctx.args.idle_timeout,
+                ctx.args.timeout,
+                reviewer_tool.as_str(),
+            );
+        let stream_mode = ctx.stream_mode;
+        let idle_timeout_seconds = ctx.idle_timeout_seconds;
+        let readonly_project_root = ctx.readonly_project_root;
+        join_set.spawn(async move {
+            let session_result = execute_review_with_tier_filter(
+                reviewer_tool,
+                reviewer_prompt,
+                None,
+                reviewer_model,
+                reviewer_model_spec,
+                reviewer_tier_name,
+                false,
+                None,
+                reviewer_thinking,
+                reviewer_description,
+                &reviewer_project_root,
+                reviewer_config.as_ref(),
+                &reviewer_global,
+                reviewer_pre_session_hook,
+                reviewer_routing,
+                stream_mode,
+                idle_timeout_seconds,
+                reviewer_initial_response_timeout_seconds,
+                reviewer_force_override,
+                reviewer_force_ignore_tier,
+                reviewer_no_failover,
+                reviewer_fast_but_more_cost,
+                false,
+                reviewer_no_fs_sandbox,
+                readonly_project_root,
+                &reviewer_extra_writable,
+                &reviewer_extra_readable,
+            )
+            .await?;
+            build_reviewer_outcome(reviewer_index, reviewer_tool, &session_result)
+        });
+    }
+
+    let outcomes =
+        collect_reviewer_outcomes(&mut join_set, &reviewer_tool_plan, ctx.args.timeout).await?;
+
+    let review_iterations = outcomes
+        .first()
+        .map(|outcome| resolve_review_iterations(ctx.project_root, &outcome.session_id))
+        .unwrap_or(1);
+    let head_sha = csa_session::detect_git_head(ctx.project_root).unwrap_or_default();
+    let diff_fingerprint = compute_diff_fingerprint(ctx.project_root, ctx.scope);
+    persist_multi_review_sidecars(
+        ctx.project_root,
+        ctx.scope,
+        &outcomes,
+        &head_sha,
+        review_iterations,
+        diff_fingerprint.clone(),
+    );
+
+    let responses: Vec<AgentResponse> = outcomes
+        .iter()
+        .map(|o| AgentResponse {
+            agent: format!("reviewer-{}:{}", o.reviewer_index + 1, o.tool.as_str()),
+            content: o.verdict.to_string(),
+            weight: 1.0,
+            timed_out: false,
+        })
+        .collect();
+
+    let consensus_result = resolve_consensus(consensus_strategy, &responses);
+    let all_reviewers_unavailable = !outcomes.is_empty()
+        && outcomes
+            .iter()
+            .all(|outcome| outcome.verdict == crate::review_consensus::UNAVAILABLE);
+    let final_verdict = if all_reviewers_unavailable {
+        crate::review_consensus::UNAVAILABLE
+    } else {
+        consensus_verdict(&consensus_result)
+    };
+    let agreement = agreement_level(&consensus_result);
+    let consensus_artifacts = super::parent_artifacts::MultiReviewerConsensusArtifacts {
+        project_root: ctx.project_root,
+        reviewers: ctx.reviewers,
+        outcomes: &outcomes,
+        final_verdict,
+        all_reviewers_unavailable,
+        head_sha: &head_sha,
+        scope: ctx.scope,
+        review_iterations,
+        diff_fingerprint: diff_fingerprint.clone(),
+    };
+    if let Err(err) =
+        super::parent_artifacts::write_multi_reviewer_consensus_artifacts(consensus_artifacts)
+    {
+        warn!(
+            error = %err,
+            "Failed to write multi-reviewer consensus artifacts (continuing)"
+        );
+    }
+
+    print_reviewer_outcomes(&outcomes);
+
+    println!(
+        "===== Consensus =====\nstrategy: {}\nconsensus_reached: {}\nagreement_level: {:.0}%\nfinal_decision: {final_verdict}\nindividual_verdicts:",
+        consensus_strategy_label(consensus_result.strategy_used),
+        consensus_result.consensus_reached,
+        agreement * 100.0,
+    );
+    for outcome in &outcomes {
+        println!(
+            "- reviewer {} ({}) => {}",
+            outcome.reviewer_index + 1,
+            outcome.tool,
+            outcome.verdict
+        );
+    }
+
+    let review_session_ids = outcomes
+        .iter()
+        .map(|outcome| outcome.session_id.clone())
+        .collect::<Vec<_>>();
+    maybe_extract_recurring_bug_class_skills(ctx.project_root, &review_session_ids);
+    Ok(if final_verdict == CLEAN { 0 } else { 1 })
+}
 
 pub(super) fn warn_if_fast_mode_has_no_codex_reviewer(
     fast_but_more_cost: bool,
